pyproject.toml
[tool.poetry]
name = "nwkfmt"
version = "0.1.1"
description = "Simple Newick tree validator and formatter"
authors = ["Cornelius Roemer <cornelius.roemer@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
nwkfmt = "nwkfmt.cli:entry_point"

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
biopython = "^1.73"
typer = ">=0.6, <1"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
